[minor] Move descendent terminator signal handler
No code change, this is in preparation for a future commit where more
than one client will use this code.
